Infinite Scroll: ensure Core's media players are initialized with each IS load, but only initialized once
r97395 added special handling for Core's ''MediaElement.js''-based players needed by incoming post sets. That changeset, however, failed to account for the situation where players are present in the initial set of posts and in subsequent sets of posts loaded by IS. This changeset ensures that all players, regardless of when they're added, are initialized.

''MediaElement.js'' doesn't like it when already-initialized players are re-initialized, which complicated this changeset somewhat. Re-initializing initialized players causes them to break completely, which is why we unbind and rebind the custom initialization.

This changeset also causes IS to pass the response object when triggering its `post-load` event, as suggested by @bradyvercher in https://github.com/Automattic/jetpack/pull/532. Doing so allows the Mejs initialization to bail early when there are no incoming players to initialize.

Fixes #6588.

Merges r97399-wpcom.