Infinite Scroll: disable for customizer previews
When previewing a theme other than the current theme, the infinite scroll
results include structure and styles for the current theme, causing the page to
become an amalgam of the two. Eventually it would be good for the Infinite
Scroll requests to include the previewed theme, but for now disabling IS in
Customizer previews is a better experience.

This same issue applies to Jetpack sites, so the fix must be applied here rather
than in wpcom-specific files.

Fixes #6795-wpcom

Merges r115743-wpcom.